+++ src/main/scala/de/sciss/synth/message/ServerCodec.scala
@@ -1,4 +1,18 @@
-package de.sciss.synth.message
+/*
+ *  ServerCodec.scala
+ *  (ScalaCollider)
+ *
+ *  Copyright (c) 2008-2014 Hanns Holger Rutz. All rights reserved.
+ *
+ *  This software is published under the GNU General Public License v2+
+ *
+ *
+ *  For further information, please contact Hanns Holger Rutz at
+ *  contact@sciss.de
+ */
+
+package de.sciss.synth
+package message
 
 import de.sciss.osc.{Packet, Message, Bundle, PacketCodec}
 import java.io.PrintStream
@@ -31,34 +45,76 @@ object ServerCodec extends PacketCodec {
   }
 
   private final val decodeSynced: (String, ByteBuffer) => Message = { (name, b) =>
-  // ",i"
+    // ",i"
     if (b.getShort() != 0x2C69) decodeFail(name)
     skipToValues(b)
-
     val id = b.getInt()
-
     Synced(id)
+  }
+
+  private final val decodeTrigger: (String, ByteBuffer) => Message = { (name, b) =>
+    // ",iif"
+    if (b.getInt() != 0x2C696966) decodeFail(name)
+    skipToValues(b)
+
+    val nodeID  = b.getInt()
+    val trig    = b.getInt()
+    val value   = b.getFloat()
+
+    Trigger(nodeID = nodeID, trig = trig, value = value)
+  }
+
+  private def decodeSetNum(b: ByteBuffer, name: String): Int = {
+    // "if" * N
+    var num = 0
+    var tt  = b.get()
+    while (tt != 0) {
+      if (tt != 0x69) decodeFail(name)
+      tt   = b.get()
+      if (tt != 0x66) decodeFail(name)
+      num += 1
+      tt   = b.get()
+    }
+    skipToAlign(b)
+    num
+  }
+
+  private def decodeSetnNum(b: ByteBuffer, name: String): Int = {
+    // ["ii" ++ "f" * M] * N
+    var num = 0
+    var tt  = b.get()
+    while (tt != 0) {
+      // "ii"
+      if (tt != 0x69) decodeFail(name)
+      tt = b.get()
+      if (tt != 0x69) decodeFail(name)
+      num += 1
+      tt   = b.get()
+      while (tt == 0x66) tt = b.get() // "f" * M
+    }
+    skipToAlign(b)
+    num
+  }
+
+  private def decodeSetPairs(b: ByteBuffer, num: Int): Vector[FillValue] = Vector.fill(num) {
+    val index = b.getInt()
+    val value = b.getFloat()
+    FillValue(index, value)
+  }
+
+  private def decodeSetnPairs(b: ByteBuffer, num: Int): Vector[(Int, Vector[Float])] = Vector.fill(num) {
+    val index   = b.getInt()
+    val num     = b.getInt()
+    val values  = Vector.fill(num)(b.getFloat())
+    (index, values)
   }
 
   private final val decodeBufferSet: (String, ByteBuffer) => Message = { (name, b) =>
     // ",i"
     if (b.getShort() != 0x2C69) decodeFail(name)
-    // "if" * N
-    var num = 0
-    var tt  = b.getShort()
-    while (tt != 0) {
-      if (tt != 0x6966) decodeFail(name)
-      num += 1
-      tt   = b.getShort()
-    }
-    skipToAlign(b)
-
+    val num   = decodeSetNum  (b, name)
     val id    = b.getInt()
-    val pairs = Vector.fill[(Int, Float)](num) {
-      val index = b.getInt()
-      val value = b.getFloat()
-      (index, value)
-    }
+    val pairs = decodeSetPairs(b, num )
 
     BufferSet(id, pairs: _*)
   }
@@ -66,28 +122,29 @@ object ServerCodec extends PacketCodec {
   private final val decodeBufferSetn: (String, ByteBuffer) => Message = { (name, b) =>
     // ",i"
     if (b.getShort() != 0x2C69) decodeFail(name)
-    // ["ii" ++ "f" * M] * N
-    var num = 0
-    var tt  = b.getShort()
-    while (tt != 0) {
-      // "ii"
-      if (tt != 0x6969) decodeFail(name)
-      num += 1
-      tt   = b.get()
-      while (tt == 0x66) tt = b.get() // "f" * M
-      if (tt == 0x69) tt = ((tt << 8) | b.get()).toShort
-    }
-    skipToAlign(b)
-
+    val num   = decodeSetnNum  (b, name)
     val id    = b.getInt()
-    val pairs = Vector.fill[(Int, Vector[Float])](num) {
-      val index   = b.getInt()
-      val num     = b.getInt()
-      val values  = Vector.fill(num)(b.getFloat())
-      (index, values)
-    }
+    val pairs = decodeSetnPairs(b, num )
 
     BufferSetn(id, pairs: _*)
+  }
+
+  private final val decodeControlBusSet: (String, ByteBuffer) => Message = { (name, b) =>
+    // ","
+    if (b.get() != 0x2C) decodeFail(name)
+    val num   = decodeSetNum  (b, name)
+    val pairs = decodeSetPairs(b, num )
+
+    ControlBusSet(pairs: _*)
+  }
+
+  private final val decodeControlBusSetn: (String, ByteBuffer) => Message = { (name, b) =>
+    // ","
+    if (b.get() != 0x2C) decodeFail(name)
+    val num   = decodeSetnNum  (b, name)
+    val pairs = decodeSetnPairs(b, num )
+
+    ControlBusSetn(pairs: _*)
   }
 
   private def decodeNodeChange(factory: NodeMessageFactory): (String, ByteBuffer) => Message = { (name, b) =>
@@ -136,17 +193,20 @@ object ServerCodec extends PacketCodec {
 
   private final val msgDecoders = Map[String, (String, ByteBuffer) => Message](
     "/status.reply" -> decodeStatusReply,
-    "/n_go"         -> decodeNodeChange(NodeGo),
-    "/n_end"        -> decodeNodeChange(NodeEnd),
-    "/n_off"        -> decodeNodeChange(NodeOff),
-    "/n_on"         -> decodeNodeChange(NodeOn),
+    "/n_go"         -> decodeNodeChange(NodeGo  ),
+    "/n_end"        -> decodeNodeChange(NodeEnd ),
+    "/n_off"        -> decodeNodeChange(NodeOff ),
+    "/n_on"         -> decodeNodeChange(NodeOn  ),
     "/n_move"       -> decodeNodeChange(NodeMove),
     "/n_info"       -> decodeNodeChange(NodeInfo),
     "/synced"       -> decodeSynced,
     "/b_set"        -> decodeBufferSet,
     "/b_setn"       -> decodeBufferSetn,
+    "/c_set"        -> decodeControlBusSet,
+    "/c_setn"       -> decodeControlBusSetn,
     "/b_info"       -> decodeBufferInfo,
-    "status.reply"  -> decodeStatusReply
+    "/tr"           -> decodeTrigger,
+    "status.reply"  -> decodeStatusReply  // old SC versions dropped the slash
   )
 
   private final val superDecoder: (String, ByteBuffer) => Message =