+++ CHANGES.md
@@ -7,6 +7,7 @@ This document only highlights major changes which probably require some adjustme
 * fixes #31, #36, #40, UGens #10, more UGen docs and examples
 * more consistent argument names, relaxes `Vec[Float]` to `IndexedSeq[Float]` in some places
 * in the server configuration `programPath` was shorted to `program`
+* `ControlSetMap` was renamed to `ControlSet`; `Single` became `Value` and `Multi` became `Vector`
 * if `$SC_HOME` is not found, uses plain `"scsynth"` as program path which should correctly resolve to `$PATH` entries.
 * the server port may be left at `0` in which case `pickPort()` is called automatically.
 * the usage of callbacks (`Completion`) is replaced by futures where sensible.